api/requirements.txt
fastapi
uvicorn
pydantic
tensorflow
numpy
pickle-mixin
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
